Trying to debug and fix test data
The settings in config.yml do not match the app manes deployed in the mocked data.
Also, the number of requests per minute in the prediction is too high for
the performance given in the mocked data.

Sptd is unable to find a proprer policy, and even breaks for indexes out of range.

I'm trying to find out where is the problem.